test: move status.go tests from resources_test.go to status_test.go
Move the "PVC detection" and "PVC classification with resizing PVCs"
test blocks to status_test.go, alongside the other tests for
functions defined in status.go (EnrichStatus, classifyPVC). Keep
only the BelongToInstance and tablespace tests in resources_test.go.
