+++ pkg/reconciler/persistentvolumeclaim/resources_test.go
@@ -20,67 +20,13 @@ SPDX-License-Identifier: Apache-2.0
 package persistentvolumeclaim
 
 import (
-	batchv1 "k8s.io/api/batch/v1"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-var _ = Describe("PVC detection", func() {
-	It("will list PVCs with Jobs or Pods or which are Ready", func(ctx SpecContext) {
-		clusterName := "myCluster"
-		makeClusterPVC := func(serial string, isResizing bool) corev1.PersistentVolumeClaim {
-			return makePVC(clusterName, serial, serial, NewPgDataCalculator(), isResizing)
-		}
-		pvcs := []corev1.PersistentVolumeClaim{
-			makeClusterPVC("1", false), // has a Pod
-			makeClusterPVC("2", false), // has a Job
-			makeClusterPVC("3", true),  // resizing
-			makeClusterPVC("4", false), // dangling
-		}
-		cluster := &apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: clusterName,
-			},
-		}
-		EnrichStatus(
-			ctx,
-			cluster,
-			[]corev1.Pod{
-				makePod(clusterName, "1", specs.ClusterRoleLabelPrimary),
-				makePod(clusterName, "3", specs.ClusterRoleLabelReplica),
-			},
-			[]batchv1.Job{makeJob(clusterName, "2")},
-			pvcs,
-		)
-
-		Expect(cluster.Status.PVCCount).Should(BeEquivalentTo(4))
-		Expect(cluster.Status.InstanceNames).Should(Equal([]string{
-			clusterName + "-1",
-			clusterName + "-2",
-			clusterName + "-3",
-			clusterName + "-4",
-		}))
-		Expect(cluster.Status.InitializingPVC).Should(Equal([]string{
-			clusterName + "-2",
-		}))
-		Expect(cluster.Status.ResizingPVC).Should(Equal([]string{
-			clusterName + "-3",
-		}))
-		Expect(cluster.Status.DanglingPVC).Should(Equal([]string{
-			clusterName + "-4",
-		}))
-		Expect(cluster.Status.HealthyPVC).Should(Equal([]string{
-			clusterName + "-1",
-		}))
-		Expect(cluster.Status.UnusablePVC).Should(BeEmpty())
-	})
-})
 
 var _ = Describe("PVCs used by instance", func() {
 	clusterName := "cluster-pvc-instance"
@@ -106,53 +52,6 @@ var _ = Describe("PVCs used by instance", func() {
 	It("fails when trying to get a pvc that doesn't belong to the instance", func() {
 		res := BelongToInstance(cluster, instanceName, instanceName+"-nil")
 		Expect(res).To(BeFalse())
-	})
-})
-
-var _ = Describe("PVC classification with resizing PVCs", func() {
-	clusterName := "myCluster"
-	makeCluster := func() *apiv1.Cluster {
-		return &apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{Name: clusterName},
-		}
-	}
-
-	It("classifies resizing PVC without pod and with FileSystemResizePending as dangling", func(ctx SpecContext) {
-		pvc := makePVC(clusterName, "1", "1", NewPgDataCalculator(), true)
-		pvc.Status.Conditions = append(pvc.Status.Conditions, corev1.PersistentVolumeClaimCondition{
-			Type: corev1.PersistentVolumeClaimFileSystemResizePending, Status: corev1.ConditionTrue,
-		})
-		cluster := makeCluster()
-		EnrichStatus(ctx, cluster, []corev1.Pod{}, []batchv1.Job{}, []corev1.PersistentVolumeClaim{pvc})
-		Expect(cluster.Status.DanglingPVC).Should(Equal([]string{clusterName + "-1"}))
-		Expect(cluster.Status.ResizingPVC).Should(BeEmpty())
-	})
-
-	It("classifies resizing PVC without pod and without FileSystemResizePending as dangling", func(ctx SpecContext) {
-		pvc := makePVC(clusterName, "1", "1", NewPgDataCalculator(), true)
-		cluster := makeCluster()
-		EnrichStatus(ctx, cluster, []corev1.Pod{}, []batchv1.Job{}, []corev1.PersistentVolumeClaim{pvc})
-		Expect(cluster.Status.DanglingPVC).Should(Equal([]string{clusterName + "-1"}))
-		Expect(cluster.Status.ResizingPVC).Should(BeEmpty())
-	})
-
-	It("classifies resizing PVC as dangling when its pod was deleted during rolling update (#9786)", func(ctx SpecContext) {
-		// Simulate simultaneous storage + resource change:
-		// instance-1 has a running pod (primary), instance-2's pod was deleted
-		// for a rolling update while both PVCs are resizing.
-		pvc1 := makePVC(clusterName, "1", "1", NewPgDataCalculator(), true) // resizing, has pod
-		pvc2 := makePVC(clusterName, "2", "2", NewPgDataCalculator(), true) // resizing, pod deleted
-		cluster := makeCluster()
-		EnrichStatus(
-			ctx,
-			cluster,
-			[]corev1.Pod{makePod(clusterName, "1", specs.ClusterRoleLabelPrimary)},
-			[]batchv1.Job{},
-			[]corev1.PersistentVolumeClaim{pvc1, pvc2},
-		)
-		Expect(cluster.Status.ResizingPVC).Should(Equal([]string{clusterName + "-1"}))
-		Expect(cluster.Status.DanglingPVC).Should(Equal([]string{clusterName + "-2"}))
-		Expect(cluster.Status.Instances).Should(BeEquivalentTo(2))
 	})
 })
